requirements.txt
absl-py==1.3.0
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
boto3==1.26.41
botocore==1.29.41
contourpy==1.0.6
cvzone==1.5.6
cycler==0.11.0
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flatbuffers==22.12.6
fonttools==4.38.0
gunicorn==20.1.0
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.6.2
mediapipe==0.9.0.1
numpy==1.24.1
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.3.0
protobuf==3.20.3
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
